requirements.txt
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
boto3==1.24.75
botocore==1.27.75
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.83.0
h11==0.13.0
httptools==0.5.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.2
MarkupSafe==2.1.1
orjson==3.8.0
psycopg2
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
